Reorganize integration tests and fixtures (#689)
* Isolate integration fixtures to integration tests

* Reorganize tests
